+++ src/module/apps/acc_diff/AccDiffInput.svelte
@@ -12,7 +12,7 @@
   <i class="cci cci-accuracy i--sm" />
 </button>
 <label for={id} class="flexcol" data-tooltip="Global Accuracy/Difficulty Adjustment">
-  <strong>Manual Adjust</strong>
+  <strong style="text-wrap: nowrap">Manual Adjust</strong>
   <strong class="accdiff-value">
     <span>{Math.abs(value)}</span>
     <i class="i--sm cci" class:cci-accuracy={value >= 0} class:cci-difficulty={value < 0} />