+++ src/module/util/lid.ts
@@ -4,7 +4,7 @@ import { LancerActor, LancerDEPLOYABLE } from "../actor/lancer-actor";
 export function lookupOwnedDeployables(owner: LancerActor, filter?: string[]): Record<string, LancerDEPLOYABLE> {
   if (owner.is_deployable()) return {};
   if (owner.isToken) return {}; // This might be possible if we could recover the original actor somehow?
-  let foundDeployables = game.actors!.filter(a => !!(a.is_deployable() && a.system.owner?.value == owner));
+  let foundDeployables = game.actors!.filter(a => !!(a.is_deployable() && a.system.owner?.id === owner.uuid));
   let result: Record<string, LancerDEPLOYABLE> = {};
   for (let dep of foundDeployables as unknown as LancerDEPLOYABLE[]) {
     if (!filter || filter.includes(dep.system.lid)) {